requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
jose
requests
uuid